+++ Project/app/src/main/java/com/example/project/fragment/TodayFragment.kt
@@ -1,10 +1,13 @@
 package com.example.project.fragment
 
 import android.os.Bundle
+import android.provider.BaseColumns
 import android.view.View
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProviders
+import com.example.project.DBHelper
+import com.example.project.PhysicalRecordContract
 import com.example.project.R
 import com.example.project.viewmodel.SharedViewModel
 import java.time.LocalDateTime
@@ -46,6 +49,30 @@ class TodayFragment : Fragment(R.layout.fragment_today) {
 
         // viewmodelの値をtextViewに格納
         updateView()
+
+        val dbHelper = DBHelper(activity!!)
+
+        val db = dbHelper.readableDatabase
+
+        val projection = arrayOf(
+            BaseColumns._ID,
+            PhysicalRecordContract.PhysicalRecordEntry.COLUMN_NAME_BODY_WEIGHT,
+            PhysicalRecordContract.PhysicalRecordEntry.COLUMN_NAME_BODY_FAT_PERCENTAGE,
+            PhysicalRecordContract.PhysicalRecordEntry.COLUMN_NAME_CREATED_AT)
+
+        val dateBegin = model.dateToday + " 23:59:59"
+        val dateEnd = model.dateToday + " 00:00:00"
+
+        val sql = "select bodyWeight, bodyFatPercentage, createdAt from physicalRecord where createdAt <= ? and createdAt >= ?  limit 1;"
+        val cursor = db.rawQuery(sql, arrayOf(dateBegin, dateEnd))
+
+        with(cursor) {
+            while (moveToNext()) {
+                model.bodyWeight = cursor.getString(0)
+                model.bodyFatPercentage = cursor.getString(1)
+                model.basalMetabolicRate = cursor.getString(2)
+            }
+        }
     }
 
     private fun updateView() {
