+++ Project/app/src/main/java/com/example/project/viewmodel/SharedViewModel.kt
@@ -1,6 +1,5 @@
 package com.example.project.viewmodel
 
-import android.graphics.Bitmap
 import androidx.lifecycle.ViewModel
 
 class SharedViewModel : ViewModel() {
